check if window and navigator exist
In cases where this package is imported in a Server Sided Rendering context, there is no window object and also no navigator. I'd like to propose checks for window because builds would fail on `"navigator" is not available during server side rendering.`

With these checks the code is only executed when actually inside a browser.